Updates
1. Map panel will listen for key events to move around the view
2. Map panel will render the secondary map sample, for placement purposes
3. Map panel's cell size is better controller now
4. Map Sample Area has more method to make working with it easier, but it could be sped up

Need to work on the map builder panel, which will work with the map panel in order to grab the required map chunks from the device, convert them into a sample and manage them.  Maybe a sample needs to be keep in memory and applied later when we get better alignment.  Also need a way to convert a floor to a sample, so we can build them out in larger chunks.